コンポーネントをファイルに分割
