+++ src/Config.js
@@ -1,0 +1,9 @@
+// アプリ全体で使用する設定値を定義する
+
+// 1〜の値に対して記号を付ける
+export const presetMentals = {
+  1: "😲",
+  2: "😫",
+  3: "😐",
+  4: "😀"
+}
